requirements.txt
requests>=2.28.0
pyyaml>=6.0
apscheduler>=3.10.0
